Rename io to inout
