+++ disruption_py/settings/cache_setting.py
@@ -8,7 +8,7 @@ from typing import Dict, Union
 import pandas as pd
 
 from disruption_py.core.utils.enums import map_string_to_enum
-from disruption_py.io.sql import ShotDatabase
+from disruption_py.inout.sql import ShotDatabase
 from disruption_py.machine.tokamak import Tokamak
