+++ disruption_py/settings/nickname_setting.py
@@ -6,8 +6,8 @@ from logging import Logger
 from typing import Dict, Union
 
 from disruption_py.core.utils.enums import map_string_to_enum
-from disruption_py.io.mds import MDSConnection
-from disruption_py.io.sql import ShotDatabase
+from disruption_py.inout.mds import MDSConnection
+from disruption_py.inout.sql import ShotDatabase
 from disruption_py.machine.tokamak import Tokamak
 
 NicknameSettingType = Union[
