+++ disruption_py/settings/time_setting.py
@@ -11,8 +11,8 @@ import pandas as pd
 
 from disruption_py.config import config
 from disruption_py.core.utils.enums import map_string_to_enum
-from disruption_py.io.mds import MDSConnection
-from disruption_py.io.sql import ShotDatabase
+from disruption_py.inout.mds import MDSConnection
+from disruption_py.inout.sql import ShotDatabase
 from disruption_py.machine.tokamak import Tokamak
