+++ docs/usage/mds_connection_reference.md
@@ -1,7 +1,7 @@
 
-The [`MDSConnection`][disruption_py.io.mds.MDSConnection] class should be used for all data retrieval tasks from MDSplus. It is a simple wrapper for the MDSplus thin client. 
+The [`MDSConnection`][disruption_py.inout.mds.MDSConnection] class should be used for all data retrieval tasks from MDSplus. It is a simple wrapper for the MDSplus thin client. 
 
-::: disruption_py.io.mds
+::: disruption_py.inout.mds
     handler: python
     options:
         filters: ["!^_[^_]"]
