+++ drafts/scripts/profile_shots.py
@@ -13,7 +13,7 @@ import pstats
 import time
 from pstats import SortKey
 
-from disruption_py.io.sql import create_cmod_handler, create_d3d_handler
+from disruption_py.inout.sql import create_cmod_handler, create_d3d_handler
 from disruption_py.machine.factory import get_shot_class_for_shot_id
 from disruption_py.shots import CModShot, D3DShot
